remove TextNodePrivate
